+++ server/user.go
@@ -2,16 +2,16 @@ package server
 
 type User struct {
 	Address
-	userID int
+	userID int64
 }
 
-func NewUser(address Address, userID int) *User {
+func NewUser(address Address, userID int64) *User {
 	return &User{
 		Address: address,
 		userID:  userID,
 	}
 }
 
-func (u *User) UserID() int {
+func (u *User) UserID() int64 {
 	return u.userID
 }
